- lophttpd: better performance on heavy load (c10k):   dynamically re-calculating the sendfile() chunksize to not suffer   from connection drops due to overruning TCP send buffers

+++ config.h
@@ -19,7 +19,7 @@ namespace httpd_config
 	extern uid_t user_uid, user_gid;
 	extern std::string host, port;
 	extern int cores, master, af;
-	extern size_t mss;
+	extern uint16_t mss;
 }
